--- a/src/Impute.jl
+++ b/src/Impute.jl
@@ -70,10 +70,10 @@ for (f, v) in pairs(imputation_methods)
     f! = Symbol(f, :!)
 
     @eval begin
-        $f(data; kwargs...) = impute($typename(; _extract_context_kwargs(kwargs...)...), data)
-        $f!(data; kwargs...) = impute!($typename(; _extract_context_kwargs(kwargs...)...), data)
-        $f(; kwargs...) = data -> impute($typename(; _extract_context_kwargs(kwargs...)...), data)
-        $f!(; kwargs...) = data -> impute!($typename(; _extract_context_kwargs(kwargs...)...), data)
+        $f(data; kwargs...) = impute(data, $typename(; _extract_context_kwargs(kwargs...)...))
+        $f!(data; kwargs...) = impute!(data, $typename(; _extract_context_kwargs(kwargs...)...))
+        $f(; kwargs...) = data -> impute(data, $typename(; _extract_context_kwargs(kwargs...)...))
+        $f!(; kwargs...) = data -> impute!(data, $typename(; _extract_context_kwargs(kwargs...)...))
     end
 end