@nalimilan 2019-07-09
> I think we probably should support taking the `obsdim` as a kwarg.
> It is what `StatsBase` is moving towards if i recall discussions with @nalimilan
> correctly

Unfortunately I don't think we have clearly agreed on the standard keyword argument for this. `pairwise` in Distances uses `dims`, like `cov` and co. in Statistics, but that's not super obvious. So we could use `obsdim` or `vardim`, but we need to discuss that somewhere. One issue is that for `pairwise` the question isn't really where are observations and where are variables, but rather what you want to compute (both could make sense). But maybe that's a special function and in most cases `obsdim` or `vardim` is OK.