@oxinabox 2019-07-09
but we have at least settled that where possible it should be a kwarg
and not just baked into the function itself. (and mentioned in doc string).