@nickrobinson251 2019-07-09
fwiw [CovarianceEstimation.jl](https://github.com/mateuszbaran/CovarianceEstimation.jl) also takes a `dims` kwarg for specifying observations