@rofinn 2019-07-09
That's true, but the annoying thing is that `dims` only makes sense when the input is a matrix. I'd also rather not need to pass kwargs to the internal `impute` functions because our handy functions extract the kwargs into an imputor type internally.