@rofinn 2019-07-09
I talked to Curtis and his suggestion was to always include a `dim` keyword to the `Imputor` constructors and only use it when the input is a matrix. That's still pretty ugly, but it maintains the relatively simple internal API.