@rofinn 2019-07-09
Based on a recommendation from Curtis I've introduced a `vardim` kwarg to the `Imputor` constructors and imputation convenience functions. The `vardim` field in each Imputor is ignored unless we're operating on matrices.