@oxinabox 2019-07-09
> Based on a recommendation from Curtis I've introduced a vardim kwarg to the Imputor constructors and imputation convenience functions. The vardim field in each Imputor is ignored unless we're operating on matrices.

Just for my edification why `vardim` and not `obsdim`?
I always think in terms of `obsdim.
(as I said though I don't feel it matters one way or the other as lomg as it is configurable)
I guess `vardim` because most imputation treat each variable as indipendeng?
so that is the one that matters?
(but on that basis it is also best for each var to stay as a column since that will be most cache friendly)