@rofinn 2019-07-09
> I guess vardim because most imputation treat each variable as independent?

More because it's already been used in the stats ecosystem and it reads better to me than `obsdim` 🤷‍♂ ?