@rofinn 2019-07-10
Alright, I think this is sufficiently better that it warrants another review at some point. Apart from the requested changes from @oxinabox (thanks for review this massive thing!) most of the other changes were just to update the docs to be more consistent and more examples focused.