@rofinn 2019-07-11
> API-wise i slightly worry there are too many different ways to do things, and I personally get very confused by |> syntax. I would be very happy with only impute!(::Imputor, data) and will be sad if I need any other syntax.

The main issue with `impute!(::Imputor, data)` is that it's verbose and the primary reason we're using types is to provide reasonable fallbacks for methods for matrices and tables in the univariate cases.