@rofinn 2019-07-11
Since folks seem largely against it I'm just going to remove the matrix orientation deprecation. Any methods using Distributions.jl or MultivariateStats.jl can just use the `vardim` field to determine if we need to pivot the matrix.