@rofinn 2019-07-11
Okay, I've resolved and applied most of the recommendations. I've commented on things that should be handled in a separate PR and responded if I disagree (e.g., I think `|>` makes sense here and you can always use explicit function calls if you don't like that).