@rofinn 2019-07-15
Alright, since I don't have as strong of an argument for using a pipe to the `materializer` I've opted to change that. I still have a strong preference for using julia's `|>` and `∘` operators for composing imputation pipelines though.